fix: Refresh log listing when the workspace log dir changes (#8)
* fix: Refresh log listing when the workspace log dir changes

We were watching the environment file for specific changes (made directly to the file), but if the Logging Configuration activity panel was used to edit the log directory, no environmentChanged event fires (since the change is being made by the environment context itself).

This will trigger a refresh remotely for this case.

* update lock file

---------

Co-authored-by: Charles Teague <charles@merdianlabs.ai>